feat(safrole): entropy accumulation and key rotations (#10)
* feat(testing): run tests for safrole

* docs(safrole): notes for entropy accumulation

* feat(safrole): calculate eta_0 on epoch change

* feat(safrole): gather eta transition to a new function

* feat(safrole): generate makers on epoch change

* feat(crypto): introduce vrf from bandersnatch-vrfs-spec

* feat(crypto): calculate bandersnatch key root

* test(safrole): assert each of state field to detect errors

* docs(dev): letters in Greek alphabet

* chore(typos): introduce typos config

+++ crates/core/src/lib.rs
@@ -6,7 +6,7 @@ pub mod work;
 
 pub const VALIDATORS_COUNT: u16 = 6;
 pub const CORES_COUNT: u16 = 2;
-pub const EPOCH_LENGTH: u16 = 12;
+pub const EPOCH_LENGTH: u32 = 12;
 pub const MAX_BLOCKS_HISTORY: u16 = 8;
 pub const VALIDATORS_SUPER_MAJORITY: u16 = 5;
 pub const AVAIL_BITFIELD_BYTES: u16 = 1;
